created separate sub component for detected headers and added mock data file to gitignore
